+++ common/src/main/kotlin/computer/obscure/piku/common/ui/components/Text.kt
@@ -9,9 +9,9 @@ open class Text(
 ) : Component() {
     override val compType: CompType = CompType.TEXT
 
-    override fun width(): Float {
-        val computed = computedSize?.x ?: 0f
-        if (computed > 0f) return computed
+    override fun width(): Double {
+        val computed = computedSize?.x ?: 0.0
+        if (computed > 0.0) return computed
 
         val lines = props.text.split("\n")
         val widest = lines.maxOfOrNull { it.mcWidth() } ?: 0
@@ -19,9 +19,9 @@ open class Text(
         return (baseWidth * props.scale.x)
     }
 
-    override fun height(): Float {
-        val computed = computedSize?.y ?: 0f
-        if (computed > 0f) return computed
+    override fun height(): Double {
+        val computed = computedSize?.y ?: 0.0
+        if (computed > 0.0) return computed
 
         val lineCount = props.text.count { it == '\n' } + 1
         val baseHeight = (7 * lineCount) + props.padding.top + props.padding.bottom
