+++ common/src/main/kotlin/computer/obscure/piku/common/ui/components/props/FlowProps.kt
@@ -3,6 +3,6 @@ package computer.obscure.piku.common.ui.components.props
 import computer.obscure.piku.common.ui.classes.FlowDirection
 
 open class FlowProps(
-    var gap: Float = 0f,
+    var gap: Double = 0.0,
     var direction: FlowDirection = FlowDirection.HORIZONTAL
 ) : CollectionProps()
